requirements.txt
django
djangorestframework
PyJWT
datetime
django-cors-headers